Move azure related reporting into corresponding package.
