+++ service/src/main/java/bio/terra/profile/service/spendreporting/azure/model/SpendCategoryType.java
@@ -1,4 +1,4 @@
-package bio.terra.profile.service.spendreporting.model;
+package bio.terra.profile.service.spendreporting.azure.model;
 
 public enum SpendCategoryType {
   COMPUTE,
