+++ service/src/main/java/bio/terra/profile/service/spendreporting/azure/model/SpendDataItem.java
@@ -1,4 +1,4 @@
-package bio.terra.profile.service.spendreporting.model;
+package bio.terra.profile.service.spendreporting.azure.model;
 
 public record SpendDataItem(
     String resourceType, Double cost, String currency, SpendCategoryType spendCategoryType) {}
